% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/twostep.R
\docType{methods}
\name{spss.predict}
\alias{spss.predict}
\title{Makes predictions from a Two-Step Cluster model produced by clustering.twostep(),
 similarly to R's predict().}
\usage{
spss.predict(object, ...)

\S4method{spss.predict}{TwoStepModel}(object, newData)
}
\arguments{
\item{object}{A fitted MLlib model}

\item{newData}{a SparkDataFrame for testing.}
}
\value{
returns a DataFrame containing predicted column
}
\description{
Makes predictions from a Two-Step Cluster model produced by clustering.twostep(),
 similarly to R's predict().
}

