% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/twostep.R
\docType{methods}
\name{spss.summary}
\alias{spss.summary}
\alias{spss.summary,TwoStepModel-method}
\title{Two-Step Cluster Summary}
\usage{
spss.summary(object)

\S4method{spss.summary}{TwoStepModel}(object)
}
\arguments{
\item{object}{a fitted ML model object.}

\item{object}{A fitted MLlib model}

\item{newData}{a SparkDataFrame for testing.}
}
\value{
a List with Two Step Model summary.
}
\description{
Two-Step Cluster Summary
}

