% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{spss.util.parameterSwitch}
\alias{spss.util.parameterSwitch}
\title{Automatic switch input value to required format
To DO ...
 1. switch for scala list?
 2. switch for other scala format}
\usage{
spss.util.parameterSwitch(parameterMap, parameterName, parameterValue)
}
\description{
Automatic switch input value to required format
To DO ...
 1. switch for scala list?
 2. switch for other scala format
}

