% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twostep.R
\docType{class}
\name{TwoStepModel-class}
\alias{TwoStepModel-class}
\title{S4 class that represents an TwoStep Model}
\arguments{
\item{jobj}{a Java object reference to the Wrapper}
}
\description{
S4 class that represents an TwoStep Model
}
\note{
TwoStep for SPSS on SPARK
}

